condor

+++ utils/graphs.py
@@ -107,9 +107,9 @@ def ABCD(n = 1000, deg_exp = 2.16, com_exp = 1.5, seed = 10, xi = 0.2):
 
     G = nx.Graph(ig.Graph.Read_Ncol('net.dat', directed=False).get_edgelist())
 
-    os.remove('comm.dat')
-    os.remove('cs.dat')
-    os.remove('deg.dat')
+    os.remove('comm.dat') if os.path.exists('comm.dat') else None
+    os.remove('cs.dat') if os.path.exists('cs.dat') else None
+    os.remove('deg.dat') if os.path.exists('deg.dat') else None
 
     G.name = 'ABCD'
     return G
